refactor: update import paths for useToast to use hooks directory
